Table added

+++ EmployeeManagementSystem/EmployeeManagementSystem/ReactClient/employee-app/src/components/EmployeeComponent.js
@@ -1,10 +1,18 @@
-
 import React, { useState, Fragment, useEffect } from 'react'
-import Table from '../common/Table';
+import EmployeeTable from '../common/EmployeeTable';
 import AddEmployee from './AddEmployee';
 import EditEmployee from './EditEmployee';
 import EmployeeService from "../services/EmployeeService";
+import MuiAlert from '@material-ui/lab/Alert';
+import Snackbar from '@material-ui/core/Snackbar';
+
+function Alert(props) {
+    return <MuiAlert elevation={6} variant="filled" {...props} />;
+}
+
 const EmployeeComponent = () => {
+    const [open, setOpen] = React.useState(false);
+    const [msg, setMsg] = React.useState("");
     const [employees, setEmployees] = useState([]);
     const initialFormState = {
         id: null,
@@ -15,6 +23,16 @@ const EmployeeComponent = () => {
         currentAddress: '',
         ctc: 0
     }
+    const handleClick = () => {
+        setOpen(true);
+      };
+      const handleClose = (event, reason) => {
+        if (reason === 'clickaway') {
+          return;
+        }
+    
+        setOpen(false);
+      };
     // Data
     useEffect(() => {
         getEmployeeData();
@@ -32,7 +50,6 @@ const EmployeeComponent = () => {
     };
 
     // Setting state
-
     const [currentEmployee, setCurrentEmployee] = useState(initialFormState)
     const [editing, setEditing] = useState(false)
 
@@ -41,26 +58,30 @@ const EmployeeComponent = () => {
         EmployeeService.createEmployee(user)
             .then(response => {
                 getEmployeeData();
+                handleClick();
+                setMsg("Employee added successfully!");
                 console.log(response.data);
             })
             .catch(e => {
                 console.log(e);
             });
-
         console.log("Add Employee", user);
-        //setEmployees([ ...users, user ])
     }
 
     const deleteEmployee = id => {
-        EmployeeService.deleteEmployee(id)
-            .then(response => {
-                setEditing(false);
-                getEmployeeData();
-                console.log(response.data);
-            })
-            .catch(e => {
-                console.log(e);
-            });
+        if (window.confirm("Do you want to delete employee?") == true) {
+            EmployeeService.deleteEmployee(id)
+                .then(response => {
+                    setEditing(false);
+                    getEmployeeData();
+                    handleClick();
+                    setMsg("Employee deleted successfully!");
+                    console.log(response.data);
+                })
+                .catch(e => {
+                    console.log(e);
+                });
+        }
     }
 
     const updateEmployee = (id, updatedEmployee) => {
@@ -68,6 +89,8 @@ const EmployeeComponent = () => {
             .then(response => {
                 setEditing(false);
                 getEmployeeData();
+                handleClick();
+                setMsg("Employee updated successfully!");
                 console.log(response.data);
             })
             .catch(e => {
@@ -110,8 +133,13 @@ const EmployeeComponent = () => {
             </div>
             <div className="flex-large">
                 <h2>View Employee</h2>
-                <Table employees={employees} editRow={editRow} deleteEmployee={deleteEmployee} />
+                <EmployeeTable employees={employees} editRow={editRow} deleteEmployee={deleteEmployee} />
             </div>
+            <Snackbar open={open} autoHideDuration={6000} onClose={handleClose}>
+        <Alert onClose={handleClose} severity="success">
+          {msg}
+        </Alert>
+      </Snackbar>
         </div>
     )
 }
